wfe: don't log InternalErrors on 404 (#6702)
For simple 404s, there's no need to log an InternalError in addition to
the user-facing error, so pass `nil` to sendError as the internalError
parameter. This cleans up Certificate, GetOrder, and FinalizeOrder; all
the places I could find that checked for `NotFound` and also logged an
unnecessary InternalError.

I also removed a redundant an unnecessary error wrapping, and a
reference to "short serial" which is not a concept we have anymore.