Conditionally call Sprintf in log package
The Debug, Info, Warn, and Err log levels are now helper functions that call their respective <LogLevel>f function.
Each <LogLevel>f function takes a variadic amount of format specifiers to pass to auditAtLevel or logAtLevel which in turn
will perform a nil check on the variadic input parameter 'a' and call/ignore an Sprintf accordingly.