issuance: split linting and issuing
It's useful to be able to get a copy of the linting certificate out of
the process, so we can store it in the database for use by processes that
want a certificate-shaped object (for instance, scanning possibly-issued
public keys for a newly discovered weakness in key generation).

To do this, while still ensuring that it's impossible to issue a
certificate without linting it, return an IssuanceToken from the linting
process. The IssuanceToken has a private field containing the template
that was used for linting. To issue a final certificate, the
IssuanceToken has to be redeemed, and only the template stored in it can
be used.