Clone the http request to add a context
This adds some commentary explaining what's going on with the
ContextInjectingRoundTripper.  During my review of the RoundTripper interface,
I noted a comment saying that RoundTripper shouldn't modify the request (to
allow reuse of requests).  While that's not going to be a problem for this
specific test, it's just as easy to do this properly and clone the request.