Upgrade Opentelemetry dependencies
This upgrades otel to v1.15.0, and the /contrib/ packages to v0.41.0
Several dependencies are upgraded as dependencies, notably grpc.

This contains a change to grpc, only mapping some grpc.Errors into span errors
if it's Unknown, DeadlineExceeded, Unimplemented, Internal, Unavailable, or
DataLoss, which should be helpful for us as we use grpc errors semantically in
Boulder, especially NotFound.