Rename CA's "--ca-addr" flag to "--addr"
Most boulder components have a command line flag to override what gRPC and
debug port they listen on, which is used in tests to run multiple instances
with the same configuration.

However, CA's flag is named "--ca-addr", and not "--addr".
This is inconsistent with SA, RA, VA, nonce, publisher, and purger.

This flag isn't used in production, where we set it in the config file.
So it shouldn't be a breaking change to rename it.