admin: add clear-email subcommand
When a user wants their email address deleted from the database but no
longer has access to their account, this allows an administrator to
clear it.

This adds `admin` as an alias for `admin-revoker`, because we'd like
the clear-email sub-command to be a part of that overall tool, but it's
not really revocation related.