ocsp-responder: reduce spaminess of "not found"
The "no response found for request" log message is pretty useless. There
are a lot of reasons why we might have no response for an OCSP request;
the most common being that someone is requesting OCSP status for a
certificate even after it has expired.

Instead of logging all "not found" cases, log only a narrower set of
cases: those where the serial doesn't exist in the `serials` table.
Since we keep the serials table long term, even expired certificates
will be in it (so long as those certificates were issued after we
started writing to the `serials` table, in late 2019).