Prevent serial prefixes from beginning with a 1
Change the max value of the CA's `SerialPrefix` config value from 255 (a byte of all 1s) to 127 (a byte of one 0 followed by seven 1s). This prevents the serial prefix from ever beginning with a 1. This is important because serials are interpreted as signed (twos-complement) integers, and are required to be positive -- a serial whose first bit is 1 is considered to be negative and therefore in violation of RFC 5280. The go stdlib fixes this for us by prepending a zero byte to any serial that begins with a 1 bit, but we'd prefer all our serials to be the same length. Restricting the prefix to be maximum 127 means we'll never accidentally use a prefix which begins with a 1.