t.sh: rebuild local boulder-tools docker container
It turns out that, even with a "build:" stanza in docker-compose.yml, the various docker compose commands won't automatically rebuild the container image unless no such image exists in the local cache. To encourage docker compose to rebuild the image when there are changes to the dockerfile, add the --build flag to our docker compose run command.