Set MaxIdleConnsPerHost to MaxIdleConns (#7349)
MaxIdleConns defaults to 100, but MaxIdleConnsPerHost is only 2. Because
there's a new Transport per client, we should allow a single host to use
all the idle connections.

A CT Log operator has notified us we're churning a lot of connections.
This should significantly help.

We might want to make the value configurable, instead of just using the
default value, but that can always come later if needed.