Check if JSON configs are properly formatted by default
This adds a new --write flag which will write out the formatted JSON files.

By default this command now checks if the files are properly formatted and
prints a list of unformatted files.

This avoids the problem of lints failing if there are uncommited changes, and
decouples this check from git.