Run config/sql change action when ready for review
Rather than running the "check-changes" action immediately when a PR is opened, wait for the PR to be ready for review (either by being taken out of draft mode, or by having reviewers added). This way it's possible to create a draft PR, then create a corresponding deployment ticket which references the PR, add that ticket to the PR description, and mark the PR ready for review, thereby pre-empting the bot so it doesn't have to leave a comment.